backend/services/prm-service/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings

# Database
sqlalchemy
alembic
psycopg2-binary
asyncpg

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart
cryptography

# HTTP Client
httpx
aiohttp

# Caching & Queue
redis
hiredis
celery
confluent-kafka

# AI & LLM (Unpinned to prevent dependency hell)
langgraph
langchain
langchain-openai
langchain-anthropic
openai
anthropic

# FHIR (Must be >=7.1.0 for Pydantic v2 support)
fhir.resources>=7.1.0
fhirclient

# Healthcare Standards
hl7apy
pydicom

# Data Validation & Processing
phonenumbers
email-validator
python-dateutil
pytz

# Utilities
python-dotenv
click
rich
loguru

# Testing
pytest
pytest-asyncio
pytest-cov
pytest-mock
faker
factory-boy
pytest-httpx

# Development
black
isort
flake8
mypy
pylint
pre-commit

# Monitoring & Observability
prometheus-client
opentelemetry-api
opentelemetry-sdk
sentry-sdk
elasticsearch

# ML & Analytics
scikit-learn
pandas
numpy
mlflow
